requirements.txt
pandas==1.1.5
joblib==1.0.1
tensorflow==2.1.0
numpy==1.19.5
Flask==1.1.2
matplotlib==3.2.2
PyYAML==5.4.1
scikit_learn==0.22.2.post1 
Keras==2.4.3
h5py==2.10.0
